+++ particle-services/src/lib.rs
@@ -43,4 +43,5 @@ mod config;
 
 pub use app_services::ServiceInfo;
 pub use config::ParticleAppServicesConfig;
+pub use config::WasmBackendConfig;
 pub use types::peer_scope::PeerScope;
